requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
ast-comments==1.3.0
astroid==4.0.4
Authlib==1.7.0
bandit==1.9.4
black==26.3.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.5
cryptography==47.0.0
cycler==0.12.1
dataclasses-json==0.6.7
dill==0.4.1
distro==1.9.0
docstring_parser==0.18.0
dparse==0.6.4
filelock==3.29.0
flake8==7.3.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
isort==8.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
joserfc==1.6.4
kiwisolver==1.5.0
libcst==1.8.6
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.9
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==2.32.0
packaging==26.2
pandas==3.0.2
parso==0.8.6
pathspec==1.1.0
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyflakes==3.4.0
PyGithub==2.9.1
Pygments==2.20.0
PyJWT==2.12.1
pylint==4.0.5
PyNaCl==1.6.2
pyparsing==3.3.2
pytest==9.0.3
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
radon==6.0.1
regex==2026.4.4
requests==2.33.1
rich==15.0.0
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stevedore==5.7.0
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
